@@ -13,16 +13,21 @@
     "deploy": "gh-pages -d web-build"
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.56",
+    "@emotion/react": "^11.4.1",
+    "@emotion/styled": "^11.3.0",
+    "@mui/icons-material": "^5.0.3",
+    "@mui/lab": "^5.0.0-alpha.50",
+    "@mui/material": "^5.0.3",
+    "@mui/styles": "^5.0.1",
+    "@mui/x-data-grid": "^5.0.0-beta.3",
+    "@mui/x-data-grid-generator": "^5.0.0-beta.3",
     "@solid/react": "^1.10.0",
     "axios": "^0.20.0",
     "chart.js": "=2.9.4",
     "chartjs-plugin-datalabels": "^0.7.0",
     "expo": "~39.0.0",
     "highlight.js": "^11.2.0",
-    "react": "~16.13.0",
+    "react": "^17.0.0",
     "react-chartjs-2": "^2.10.0",
     "react-dom": "~16.13.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-39.0.2.tar.gz",
