@@ -1,4 +1,4 @@
 from .user import User
 from .collection import CollectionModel, CreateCollectionModel, UpdateCollectionModel
-from .evaluation import EvaluationData, CreateEvaluationModel, EvaluationModel, EvaluationResults, EvaluationScore, UpdateEvaluationModel
+from .evaluation import PyObjectId, EvaluationData, CreateEvaluationModel, EvaluationModel, EvaluationResults, EvaluationScore, UpdateEvaluationModel
 from .assessment import AssessmentModel
